Move reducer tests to their own folder
