+++ src/__tests__/reducers/RootReducer.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@jest/globals'
 
-import { rootReducer } from '../reducers/RootReducer'
+import { rootReducer } from '../../reducers/RootReducer'
 
 test('initialState', () => {
   const actual = rootReducer(undefined, { type: 'DUMMY_ACTION_PLEASE_IGNORE' })
