+++ src/__tests__/reducers/cryptoExchangeReducer.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@jest/globals'
 
-import { cryptoExchange } from '../reducers/CryptoExchangeReducer'
+import { cryptoExchange } from '../../reducers/CryptoExchangeReducer'
 
 test('initialState', () => {
   const actual = cryptoExchange(undefined, { type: 'DUMMY_ACTION_PLEASE_IGNORE' })
